requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzlocal==5.2
urllib3==2.2.1
